requirements.txt
aiogram==2.22.2
aiohttp==3.8.3
aiosignal==1.2.0
alembic==1.8.1
aniso8601==9.0.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.1.0
Babel==2.9.1
blinker==1.4
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.5
cryptography==38.0.1
cycler==0.11.0
decorator==5.1.1
dnspython==2.2.1
email-validator==1.1.3
Flask==2.0.2
Flask-Dance==5.1.0
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-Minify==0.37
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
fonttools==4.37.4
frozenlist==1.3.1
google-auth==2.12.0
google-auth-oauthlib==0.5.3
gspread==5.5.0
gunicorn==20.1.0
htmlmin==0.1.12
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.9.0
infinity==1.5
intervals==0.9.2
itsdangerous==2.1.2
Jinja2==3.0.2
jsmin==3.0.1
jsonschema==4.16.0
kiwisolver==1.4.4
lesscpy==0.15.0
Mako==1.2.3
MarkupSafe==2.1.1
matplotlib==3.6.1
multidict==6.0.2
mysqlclient==2.1.1
numpy==1.23.3
oauthlib==3.2.1
packaging==21.3
pandas==1.5.0
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
ply==3.11
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python-shell==1.0.4
pytz==2022.4
rcssmin==1.1.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.29
SQLAlchemy-Utils==0.38.3
urllib3==1.26.12
URLObject==2.4.3
validators==0.20.0
Werkzeug==2.0.3
WTForms==3.0.0
WTForms-Alchemy==0.18.0
WTForms-Components==0.10.5
xxhash==3.0.0
yarl==1.8.1
zipp==3.8.1
